app/requirements.txt
flask==2.3.3
psycopg2-binary==2.9.9
google-cloud-secret-manager==2.20.0